refactor(gateway): extract transport layer into gateway/transport/ (#321)
Move protocol-specific code (SSE parsing/formatting, HTTP client pool,
ProviderInfo/KeyRing/auth) out of proxy.py and providers.py into a
dedicated gateway/transport/ subpackage. This creates a clean module
boundary for future protocol backends (gRPC, WebSocket).

New modules:
- transport/sse.py — SSE line parsing, done detection, per-provider formatters
- transport/provider_info.py — ProviderInfo, KeyRing, auth header builders
- transport/client.py — HTTP client pool, prepare_upstream, close_clients
- transport/__init__.py — public re-exports

proxy.py retains: pipeline handlers, metadata store, shim transforms,
error helpers. providers.py retains: registry, factory, default lookups.
